pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "hypo-agent"
version = "0.1.0"
description = "Single-user personal AI assistant"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Hypo-Agent Team" }
]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "structlog>=24.0.0,<26.0.0",
  "httpx>=0.28.0,<1.0.0",
  "apscheduler>=3.10.0,<4.0.0",
  "litellm>=1.60.0,<2.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "sqlite-vec>=0.1.6,<0.2.0",
  "pymupdf>=1.26.0,<2.0.0",
  "python-pptx>=1.0.0,<2.0.0",
  "python-docx>=1.1.0,<2.0.0",
  "websockets>=16.0,<17.0",
  "tiktoken>=0.12.0,<1.0.0",
  "tavily-python>=0.7.0,<1.0.0",
  "notion-client>=2.0.0",
  "python-multipart>=0.0.22",
  "playwright>=1.54.0,<2.0.0",
  "qrcode>=8.0.0,<9.0.0",
  "cryptography>=44.0.0,<46.0.0",
  "pynacl>=1.5.0,<2.0.0",
  "lark-oapi>=1.4.19,<2.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/hypo_agent"]
include = ["src/hypo_agent/templates/*.html"]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"
markers = [
  "unit: pure unit tests without external I/O or TestClient lifecycles",
  "integration: tests that exercise TestClient, in-process services, or multi-component flows",
  "cli: CLI wrapper tests",
  "slow: tests expected to take more than 5 seconds",
]
